Fix issues around using IE never started
Invoke-WebRequest uses IE to load the Dom with the HTTP response. This doesn't work when the IE instance on the server has never be opened. This is because IE shows interactive dialogs which is not supported in PowerShell.

Changed Invoke-RestMethod to Invoke-WebRequest.

Fixed logging so that non-Response exceptions get logged correctly.

Closes #33